pyproject.toml
[project]
name = "llamaindex-workflows"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "llama-index-indices-managed-llama-cloud (>=0.6.11,<0.7.0)",
    "llama-index-llms-openai (>=0.3.38,<0.4.0)",
    "llama-index-core (>=0.12.33.post1,<0.13.0)",
    "llama-index-utils-workflow (>=0.3.1,<0.4.0)"
]

[tool.poetry]
packages = [{include = "llamaindex_workflows", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
